ww8export: Fix testTableKeep
The tables in the test .ODT contain a cell with text "With merged cells".
The text is supposed to wrap, so that cell height includes the two lines.
The width of the cell is 32.72 mm. When it is exported to .DOC, due to
rounding errors, the width becomes 32.76 mm. It happens that the text
width is too close to these numbers, so depending on rounding when text
layouting, it may start to not wrap in .DOC.

This just adds an extra space character to the line, to stop depend on
rounding errors.

This also reverts commits 1aaad03294444e9e1d47348d624978218399b6af and
63cdac7a421d34de2d5f09cca40296bf01675a89, which had workarounded this
issue.

Change-Id: Ifdb7a042f78a1a3cce3a385e9db865a1b8c7b134
Reviewed-on: https://gerrit.libreoffice.org/53906
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>