tdf#117282 Do not set property 'Name' at ControlShape
Formally property 'Name' was set at com.sun.star.drawing.ControlShape,
but this Shape does not (and never did) support that Property. It
should have thrown an UnknownPropertyException in the past, but did
not do that by error. It was just ignored in SvxShape::_setPropertyValue
and lost.
Solution is to just not set it anymore. This loses nothing (was never
set, see above), but avoids the (now correctly thrown) exception.

Change-Id: Id6bdc545862f91692213afe6fb706b98d3fce211
Reviewed-on: https://gerrit.libreoffice.org/54145
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Armin Le Grand <Armin.Le.Grand@cib.de>