tdf#117333, tdf#117325 dbahsql: exception handling
Add proper exception handling for importing HSQL database.

If there are errors during migration, the migration process should
continue anyway. The first error should be shown to the user. All the
others can be seen as a warning.

Popping up several error messages would be just annoying for the user.

Change-Id: Ia726ad777fd798f064a8fde1c0062c5b05fe59d0
Reviewed-on: https://gerrit.libreoffice.org/54396
Reviewed-by: Tamás Bunth <btomi96@gmail.com>
Tested-by: Tamás Bunth <btomi96@gmail.com>