tdf#117539 Crash after cut and paste operation of a chart stick
regression from
   commit 55b98e9137fde3ca0b732f3970032cb506ead2c7
   loplugin:useuniqueptr in SfxItemPool_Impl
where I accidentally removed some fallback code.

Note that this does not fix the underlying bug where some pool is not
correctly set up.

Also, this only fixes the bug that I introduced in the above commit,
another bug in the same area has subsequently been introduced, causing
a crash in a different area when performing this operation.

Change-Id: Iaae3b0dc82c740b498e3735273820647ef28fd01
Reviewed-on: https://gerrit.libreoffice.org/54522
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>