Really don't show empty Writer document on cancelled recovery
Commit 53706a3ddeca9fb243e161a839938b9cd4b70e6a tried to make that,
but either didn't work on all platforms, or was broken later. Then,
commit 78da69c9ba2ab3f3c0f00d1f6c3b60fa304adceb handled one special
case.

The problem is that cancelling recovery returns false. It's unclear
why the distinction should be done for the cancelled recovery case:
in any case, if the recovery has opened some documents, it will be
handled in the code that enumerates frames ("no default document if
 a document was loaded by recovery..." comment). In other cases, we
need to honour user command line parameters (selected module), and
if none is specified, do the global default, which is currently to
show Start Center. So - let the OpenDefault() do just that.

Change-Id: I90f3581fc72628d1317e97074a4a1d352d630c7e
Reviewed-on: https://gerrit.libreoffice.org/54604
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>