tdf#117247 just send a GRAPHIC_ARRIVED hint after update
As we always load graphic complete now, we can just hint that it
has arrived. In the future we might properly restore retrieving the
stream of a graphic throught link manager, but not creating the
partial graphic from it - just an async retrieval of the bytes.
The graphic should then always be created completely with the
GraphicFilter.

Change-Id: I8560c9d41a057d228aa558316131ba99d436645b
Reviewed-on: https://gerrit.libreoffice.org/54693
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>