Add FullScreen bit to WindowStateState
FullScreen in this context means a top-level window that is both
full-screen from the desktop environment's perspective (which
depending on the desktop environment might simply mean that it is as
large as possible to fill its screen, leaving any system menus etc
visible), *and* is in LibreOffice's full-screen mode (with no toolbars
or other UI elements except the document contents visible).

Not yet used, will be used in follow-up commits.

Change-Id: Ia6f86e0d2a7c5a621c6f19d897d3b17ba6bfb8b4
Reviewed-on: https://gerrit.libreoffice.org/54733
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tor Lillqvist <tml@collabora.com>