tdf#117503 DOCX import: fix out of sync first/later top margin
Improve the synchronization between the first page and follow page
style, so that when the DOC export invokes
sw::util::IsPlausableSingleWordSection(), it recognizes that originally
these were the same Word section.

If they don't match up, then the exporter will insert an explicit
section break, meaning we get one more page in the export result
compared to the original.

Change-Id: I336cf347698187cdede47be2659cec51e4381e85
Reviewed-on: https://gerrit.libreoffice.org/54965
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>