Resolves: tdf#117819 append trailing '0' as needed before separator insertion
To insert separators, literal strings between digits and other,
the formatter operates backwards on the string obtained from
doubleToUString() cleaned of the decimal separator. The number of
decimals returned by doubleToUString() may be less than the
decimals of the number format as entered by the user, which lead
to separators inserted at wrong positions. This wrong behavior was
triggered respectively made more obvious by

    commit 1197b5e3fa2bb7dbe5723bbede60b5e02bd21409
    CommitDate: Wed Jan 13 14:47:57 2016 +0100

        tdf#96918 display accurate integer double values up to (2^53)-1

which for integer values returns at most 15 decimals ('0' in this
case).

Before operating on the decimals' part ensure that the number of
decimals matches the digits requested, and if shorter append
trailing '0' characters for the required amount so the separators
(and other strings) are inserted at the correct positions.

Change-Id: Ic02652699ea7d6fae3b2b3348f6f7d183319e043
Reviewed-on: https://gerrit.libreoffice.org/55039
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>