sc: Make ZTEST() raise #DIV/0! when the sample stddev is 0
because it is an actual division-by-zero error and Excel does.

Call for suggestion: I would like to add a test case to
sc/qa/unit/data/functions/statistical/fods/ztest.fods, but
do not know the right manner of adding it cleanly without
messing up other parts such as styles.

Change-Id: I79b99a6b7385177e94d75bb4e00d724332b9b8bf
Reviewed-on: https://gerrit.libreoffice.org/55152
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Eike Rathke <erack@redhat.com>