DOCX import: fix relativeFrom=page, align=right handling again
sw/qa/extras/ooxmlexport/data/wpg-nested.docx has a right sidebar on the
page, which is a nested drawingML group shape. This was fixed to have a
reasonable position in commit abac66661a1e21c9669d2441be99fe77d1f48dc5
(DOCX import: fix relativeFrom=page, align=right handling, 2013-12-06)
but the testcase only asserted the document model, so it wasn't caught
that later commit dc3729f3ba0b05e8418815cdd0909eceba606133 (tdf#105143
DOCX import: enable DoNotCaptureDrawObjsOnPage layout compat option,
2017-01-06) broke it.

Fix it again by reverting the original fix and focusing only the part
that is still a problem: the margins around the shape prevent it to be
exactly at the right edge (instead there is some spacing). The margins
are not relevant in the "no wrap: in front of text" and "no wrap: behind
text" cases, the Word UI even disables the controls in those wrap cases.

Improve the relevant testcase to assert the layout instead, so next time
this breaks, we'll more likely notice it.

For each modified existing testcase, I compared the old Writer, new
Writer and Word rendering result and came to the conclusion that it's
the test that should be changed, not the code.

Change-Id: Ida29929671148db77fdd9ce8dbfb5214bb4728dd
Reviewed-on: https://gerrit.libreoffice.org/55192
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>