tdf#118113 MM: Make sure page count is up to date
Since the page layout may have changed, e.g. due to
the visibility status of hidden sections having changed
in the call to 'pWorkShell->SwViewShell::UpdateFields()',
it's necessary to recalculate the layout to ensure that
the correct page count is retrieved when calling
'pWorkShell->CalcLayout()'.

This fixes a regression introduced by the performance
optimization done in commit
08672f1521a4ab1ec1efc265287bd8c6d9342883, while
still avoiding to have to do the expensive layout for the
target document.

Change-Id: I46f1b68758e4df330c17358a8e852e8b040a9520
Reviewed-on: https://gerrit.libreoffice.org/55636
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>