tdf#104199 sd: export non-borders as noFill
...followup to commit f5dbde2359564f09bf843c83f65df89348374ad5
which made the border invisible by setting the color to COL_AUTO.

But being invisible isn't good enough because on a round-trip
we are now losing the "noFill" attribute and saving a defined
border. However, COL_AUTO is turned into white during import,
in both LO and in MSO, so round-tripping displayed a
white border instead of an invisible one.

Change-Id: If6cb513ca6e4336e49bc56a9509aede2e1937063
Reviewed-on: https://gerrit.libreoffice.org/55658
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Szymon Kłos <szymon.klos@collabora.com>