tdf#116208 vcl opengl: fix lack of updates in status bar
Commit 9c42c2ca9f84f055d7925db569518babd68c7943 (Cache text layout of
statusbar items, 2017-06-08) was a performance improvement of not laying
out the text of status bar text during each & every paint. One of these
places didn't work correctly with Chinese text, so commit
4f4e6bee51e251a809fd43866ca4b0bc084f4f2e (tdf#115353 Status bar: no
cache in settext, 2018-03-01) attempted to disable the caching at that
particular place by removing the update of the cache.

This resulted in lack of updates, e.g. during typing into Writer and
checking the word count part of the status bar.

Fix this (and at the same time keep the original problem fixed) by still
not using the layout cache and that problematic place, but instead of
just not updating the cache actually invalidate it -- probably that was
the intention in the force place already.

Change-Id: Ib7b91e5f24bc68d2f8cae2efea64340585ad9bd5
Reviewed-on: https://gerrit.libreoffice.org/56304
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins