xmloff: prevent embedding same font files with checksumming
When embedding fonts to ODF documents we can get into a situation
where the same font is embedded multiple times and this wastes
space. To prevent this we calculate the checksum of the font first
and track the fonts that are embedded. When the font has the same
checksum, the path to already embedded font is returned.

Change-Id: I087c42d9a63b1697b9b4ea985dbce2fda760a996
Reviewed-on: https://gerrit.libreoffice.org/56507
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>