Clean up ownership tracking of SalVisual's visual (in X11 Visual base class)
GCC trunk towards GCC 9 emits -Wdeprecated-copy because SalVisual has implicit
copy functions but a user-declared dtor.  The implicitly-defined copy functions
are actually used (so cannot be deleted), but in fragile interaction with the
semantics of the user-provided dtor in the SalColormap(sal_uInt16) ctor.

Changing SalVisual into a move-only class (moving the "this instance must delete
its visual member" information) doesn't work well, as SalDisplay::ScreenData is
used as a supply of SalVisual instances (but which will never be of the "this
instance must delete its visual member" variety) that are then copied into
SalColormap::m_aVisual.

As only SalColormap creates SalVisual instances of the "this instance must
delete its visual member" variety, what actually works is to track that
information in SalColormap instead of directly in SalVisual, and make
SalColormap a move-only class.

Change-Id: Ib968a38c40b660ce91981b3c7b281f4add6ece6b
Reviewed-on: https://gerrit.libreoffice.org/56579
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>