related tdf#82175 writerfilter: use docDefault, not w:default
Commit 4b3b0a7a for tdf#104713, added this w:default test to only
override style inheritance on well-formed documents (i.e. all default
values have been specified in the document. Otherwise the document
will have no defaults at all).

However, checking on w:default is not the best test of a well-formed
document. W:default had no impact on inheritance anyway, so it was a
bad test, and if we used that, shouldn't we setParent to
the default style instead of assuming it is "Standard"?
Besides, FindDefaultParaStyle() already exists, so a
separate boolean value wasn't necessary in that scenario.

So lots of reasons to tweak my old commit. It is prep work
for setting the parent to -none- instead of "Standard".

Change-Id: I0846c09775c9a86b506f13e3588aab437a9a93cc
Reviewed-on: https://gerrit.libreoffice.org/56824
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>