Limit GetNextPos() loops to range, if available, tdf#68290 follow-up
It's unnecessary to check all columns if there was a range marked
anyway. Also use GetPrintArea() only for bUnprotected and enlarge
the result a bit to cover the case where a protected form may be
empty but not have vidsual indications either.

Also adding nMoveY to nCol was nonsense all the time already, what
was actually meant is to add the direction of movement. Worked by
chance because nMoveY was always +1/-1.

Change-Id: Ic1175a607a1169f3ef0e28d0f3a91f8c526314f4
Reviewed-on: https://gerrit.libreoffice.org/56966
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins